NavBar y estilos
Se ha creado el componente NavBar que permite cambiar el tema de la pagina (Dark mode y Light mode) y se ha modificado el css de index y app para renderizar correctamente las Views.
